Update patch set 2
Patch Set 2:

Note that the PowerVR samples depend on the export ordinals to match. I extracted them using Visual Studio's dumpbin command line tool and manually added them to the .def. Could have a typo in it so keep that in mind when a function on the app side doesn't match up with the function being called on the lib side. I removed the ordinals for the extensions since extensions should always be queried through eglGetProcAddress.

Patch-set: 2
Label: Code-Review=0