Subzero: Fix lowering of the fabs() intrinsic.
The pand instruction for masking off the sign bit can operate on a register or an m128 memory location, but not a 32-bit or 64-bit memory location.  This means we need to make sure f32 and f64 operands are first loaded into a register.

BUG= none
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/1022123004