Update patch set 1
Patch Set 1:

This was the first attempt to implement ANDROID_image_native_buffer. A pointer to ANativeWindowBuffer is stored in egl::Image. Override/hack lock() and lockInternal() so that the data in ANativeWindowBuffer can be accessed directly, which I'm not really sure if it is a proper implementation. Not quite sure how a shared image is used by the texture.

Patch-set: 1