Update patch set 3
Patch Set 3: Verified-1

Added more details to the commit message and changed author to Ping-Hao. PTAL.

We should probably hold off committing this until we know for certain we're not running into the issue of having any of the libraries loaded twice. So I'm marking it Verified-1.

Patch-set: 3
Label: Verified=-1