Update patch set 2
Patch Set 2:

NDEBUG turns off tracing. See the definitions in:

src/Common/Debug.hpp

It's set by default in the Android build when we do userdebug, so we have to force it off here.

You are correct that it has the side-effect of turning on the UNIMPLEMENTED and assert macros. Those are actually fairly interesting at this stage of development.

Patch-set: 2
Label: Code-Review=0