Update patch set 5
Patch Set 5:

I'm looking into implementing gralloc to directly allocate SwiftShader EGL images. That would make this trivial to implement, instead of attempting to hook some generic chunk of memory into images and dealing with strides and color formats that are not optimal.

Ping-Hao, does that sound reasonable to you? Or do you know of any dependent component of gralloc that needs it to be implemented in a way that is less SwiftShader-specific?

Patch-set: 5
Label: Code-Review=0