Texture Storage for GLES 3.0
Added Texture Storage for both 2D and 3D textures.
A map was added to both verify if the format is
valid and to return a valid type. The type itself
is irrelevant, since there's not data to upload
at this point, but I made sure to return a valid
type nonetheless in order to prevent breaking any
current/future type checks in the code.

Change-Id: I49618c52427c86236afddacc4d57288f465bf971
Reviewed-on: https://swiftshader-review.googlesource.com/2812
Tested-by: Alexis Hétu <sugoi@google.com>
Reviewed-by: Nicolas Capens <capn@google.com>