Update patch set 2
Patch Set 2:

I'd like to keep the API's behavior between debug and release builds the same. Also, unless you put a breakpoint at the error() functions you wouldn't notice. Since the spec does explicitly say it's an error, just an undefined one that is not required to be reported, let's use the ERR() macro to loudly inform us about it, but not return an error to the application.

Patch-set: 2
Label: Code-Review=0