Update patch set 6
Patch Set 6:

It looks like we're willing to adopt this sooner than the debug changes. I rebased, resolved the merge conflict, and confirmed that J continues to build.

Booting seems to involve some addtional work, but I think we should merge.

PTAL and +2 if it's still acceptable after the rebase.

Patch-set: 6