Update patch set 3
Patch Set 3:

(1 comment)

The spec says that "If a value is so large in magnitude that it cannot be represented with the requested type, then the nearest value representable using the requested type is returned." So basically the 32-bit getInteger() should return 0x7FFFFFFF or 0x80000000 on overflow or underflow of a 64-bit value. I think that's best implemented by having the 32-bit one call the 64-bit one, instead of templatazing them as two separate functions.

Patch-set: 3
Label: Code-Review=0