Update patch set 4
Patch Set 4: Verified+1

Actually, after having tried to implement it, I realized that we do not have the array size information (only a pointer to an array), so it would be difficult to make a temporary pointer of the correct size without this information. I kept the template, but fixed the conversion from float to 64 bit int.

Patch-set: 4
Label: Verified=+1