Update patch set 1
Patch Set 1: Verified+1

This is a prototype to fix the read-after-delete error when using VBO. It sends a signal to the object saying that it is a potential common buffer (referenced in more than one place), thus it should not be deleted.

I am aware that this is a really specific solution and a generic one would be more appropriate. I will gladly take your suggestions :)

Patch-set: 1
Reviewer: Maxime Grégoire <5145@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Verified=+1