Update patch set 3
Patch Set 3:

(7 comments)

Note that the spec specifies that "If the GL is bound to a draw framebuffer object, the ith buffer listed in bufs must be COLOR_ATTACHMENTi or NONE. Specifying a buffer out of order, [...], will generate the error INVALID_OPERATION."

So the (draw) buffer index always matches the attachment index. It just allows to enable or disable them.

Patch-set: 3
Label: Code-Review=0