Update patch set 2
Patch Set 2: Code-Review-1

I think there's an issue here with the Device's scissorRect not being set to match the Context's scissor state in some cases. Context::clear() sets it through calling applyRenderTarget(), but the new clearColorBuffer() calls appear not to. And we don't want to call applyRenderTarget() for those.

Patch-set: 2
Label: Code-Review=-1