Update patch set 1
Patch Set 1:

This isn't badness. There are hardly any apps that use ES3, and those that do are all highly advanced games which have a fallback to ES2. So I really don't expect any issues with apps attempting to create an ES3 context. Plus we're working on ES3 support anyway.

I would value collecting stats on which apps try to create an ES3 context, but I'd prefer to do that in a platform-independent manner. Which brings us back to refactoring the logging as a whole. We need to plan that as a bigger task.

Patch-set: 1